requirements.txt
openai>=1.0.0
jsonschema>=4.18.0
python-docx
langchain-core
langchain-text-splitters